clear code

+++ mb.c
@@ -44,7 +44,7 @@ void mb_set_master_process_handler(void (*f)(mb_packet_s))
 
 #endif
 
-void mb_set_tx_handler(void (*f)(uint8_t *,uint8_t))
+void mb_set_tx_handler(void (*f)(uint8_t *,uint16_t))
 {
     mb.tx_handler=f;
 }
